backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
python-dotenv>=1.0.1
alembic>=1.12.1
requests>=2.32.3
python-multipart==0.0.9
pymysql>=1.1.0
cryptography>=41.0.7
gunicorn>=21.2.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
starlette>=0.40.0,<1.0.0
filelock>=3.13.1  # huggingface-hubの依存パッケージ
sqlalchemy-utils==0.41.1
